Fix test timeout HazelcastProxyFactoryTest.testProxyHazelcastInstanceClasses_ofSameVersion_areSame [HZ-1802]
The test is failing because it is blocked on
SocketInputStream.socketRead0(Native Method)

Ondrej Lukas said that sometimes maven central is not responsive and asked to change the test to distinguish failure due to socket read timeout.

Also added jars used in tests to maven dependency-plugin get section to install them to local repository. This will avoid unnecessary downloads

Fixes : https://github.com/hazelcast/hazelcast/issues/21401