Use TestHazelcastFactory in OrderedProcessingMultipleMemberTest [HZ-1827] [5.2.z]
The failure was caused by a wrong environment setup - the container
running the build wasn't isolated from others and another instance
joined the cluster:

```
23:15:31,673  WARN || - [NodeMulticastListener] hz.eager_chandrasekhar.MulticastThread - [127.0.0.1]:5702 [dev] [5.2-SNAPSHOT] New join request has been received from current master address. The UUID in the join request (cadb25fe-b130-4604-8b7c-95414c1c11ab) is different from the known master one (082dc3f5-b171-47c4-9add-a88aa5cdedb9). Suspecting the master address: [127.0.0.1]:5701
23:15:31,673  WARN || - [MembershipManager] hz.eager_chandrasekhar.MulticastThread - [127.0.0.1]:5702 [dev] [5.2-SNAPSHOT] Member [127.0.0.1]:5701 - 082dc3f5-b171-47c4-9add-a88aa5cdedb9 is suspected to be dead for reason: New join request has been received from current master address. The UUID in the join request (cadb25fe-b130-4604-8b7c-95414c1c11ab) is different from the known master one (082dc3f5-b171-47c4-9add-a88aa5cdedb9). Suspecting the master address: [127.0.0.1]:5701
```

The UUID `cadb25fe-b130-4604-8b7c-95414c1c11ab` doesn't appear anywhere
else in the log of the test or the whole build.

The test used real Hazelcast instance, which is really not needed in
this case. So this commit changes that to TestHazelcastFactory and it
will use the mock network.

Fixes #21532

Backport of #22996