Improve assert in ElasticSinkBuilderTest [HZ-936] [5.2.z]
It is not guaranteed that all cleanup logic (e.g Processor#close)
methods are called when `job.join();` returns.

Use assertTrueEventually to wait for all clients to be closed instead.
Also stop using mockito to verify that the client was closed and use
internal field via reflection instead to rule out any issue with
thread-safety in Mockito.

Fixes #19637

Backport of #23003