Introduce new types to binary compatibility test for Portable [API-1014]  (#22716)
`Decimal`, `LocalDate`, `LocalTime`, `LocalDateTime`, and
`OffsetDateTime` types were missing for Portable object on the binary
compatibility test and we are missing test cases for these types. This
PR adds these types to the Portable object in the test, and all
languages should check if their client supports these missing types for
Portable serialization and update the binary test file if they support.