Add debug logging for migrations & indexes
Tests with indexes during migrations sporadically fail.
There might be an issue with population of
indexes after migrations complete. This commit adds trace logging
to migrations and index population.
Logging is implemented in a subclass of Indexes class in test classpath,
as I wanted to avoid adding another implementation of Indexes class
in the production classpath, because it might affect JIT optimizations.