Use non-default multicast port in ClientToMemberDiscoveryTest [5.2.z] [API-1660] (#23025)
This test was failing when the servers where started with
`safe-serialization` disabled on members, but enabled on client from
time to time.

Upon investigation, I believe there are no problematic parts in the
multicast discovery code that would cause such a client to deserialize
messages sent by such members.

I believe this test could only fail if the client somehow reads a
multicast message serialized by `safe-serialization` enabled members.
This could happen because, we were using the default multicast ip and
port, and the `safe-serialization` is enabled by default. So, if we had
another test(from another test run, because this one is run in serial
mode) to send the multicast message and the client reads it, it can
discover the members in the `clientSerializationMismatchTest` test.

To probably solve this issue, I have started using a different port in
this test for multicast discovery.

Also, I saw a couple of small, problematic, but not-so-related codes
here and there and fixed them.

- The `DiscoveryService#start` was being called twice, instead of once,
in the client startup(once in the
`ClusterDiscoveryServiceBuilder` and once in the
`CandidateClusterContext`). I have eliminated the first one.
- There were a couple of fields that could have been final, and I made
them so.
- There was a variable that could have been a local variable, and I made
it so.
- There was an unnecessary object creation (around the PortValidator )
and I have eliminated it.

clean backport of https://github.com/hazelcast/hazelcast/pull/22920
this also backports this single-line change
https://github.com/hazelcast/hazelcast/pull/22895