Fix timeout failure in SqlAvroTest.test_topLevelFieldExtraction [HZ-1841]
Note : The fix is based on a guess because I was not able to reproduce it. The solution is found here
https://groups.google.com/g/confluent-platform/c/u3A7BCZEFNg

When Kafka is under load it becomes less responsive and when a new schema is to be registered the schema registry is waiting only 500 ms.
If no response is received it is throwing
io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException: Register operation timed out; error code: 50002

The code is that catches up schema registry with Kafka is in io.confluent.kafka.schemaregistry.storage.KafkaSchemaRegistry.java
I have changed the test to use a bigger timeout value to avoid it

Fixes : https://github.com/hazelcast/hazelcast/issues/23027