Use Hazelcast version in the protocol definitions
We were using the separate protocol version in the since values
of the protocol definitions, but that was causing a quite bit
of confusion.

We have decided to stick to the Hazelcast versions to make it simpler.

This PR updates the auto-generated test files with the changes
done on the protocol side and updates the since versions used in
the `LongRegister`.