Mention about the possibility of TargetNotMemberException in the PN Counter documentation [5.2.z] [API-1654] (#22981)
It is possible to get the TargetNotMemberException for client operations
when the client connects to a new cluster, but the member list is not
updated yet. Or, all the replica data members of the cluster die, but
the member list is not updated yet.

The client routes targeted invocations to random members, if a
connection to the target cannot be found, hoping that the member could
route the invocation to the correct target member.

That is problematic when the client tries to route PN counter
invocations to dead members, that it has no connection. The invocation
service will happily route that invocation to a random member, and that
member will throw TargetNotMemberException because the actual target of
the invocation is not in its member list.

That exception is fatal for targeted exceptions, so they are not
retried. Upon receiving an exception for an invocation, the PN counter
proxy tries to select a new target using the member list. If all
possible targets in the member list are retried, the last exception is
propagated to the user.

There is really not much we could do on the client side for this kind of
situation. We have to have that invoking on random connections if the
target connection is not available logic because the client can lose
connections to target members at any moment, even if they are perfectly
healthy, and we want the client to continue operating if possible in
that situation. We can't really change the way member list events are
received on the client side, there is always a possibility of a late
member list event.

Probably the best we can do is to document the behavior, and this PR
does that.

I have also updated a couple of existing documentations: Corrected a
couple
of typos, and removed the possibility of UnsupportedOperationException
as 5.x clients and members cannot connect to 3.10 clusters.

I have also updated the ClientPNCounterTest#testClusterRestart so that
it now takes care of the possibility of TargetNotMemberException.

clean backport of https://github.com/hazelcast/hazelcast/pull/22963