Fix timeout failure in SqlAvroTest.test_topLevelFieldExtraction [HZ-1841] (#23037)
Note : The fix is based on a guess because I was not able to reproduce
it. The solution is found here
https://groups.google.com/g/confluent-platform/c/u3A7BCZEFNg

When Kafka is under load it becomes less responsive and when a new
schema is to be registered the schema registry is waiting only 500 ms.
If no response is received it is throwing

**io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException:
Register operation timed out; error code: 50002**

The code that catches up schema registry with Kafka is in
**io.confluent.kafka.schemaregistry.storage.KafkaSchemaRegistry.java**
That code is using **kafkastore.timeout.ms** value
I have changed the test to use a bigger timeout value to avoid timeout
failures

Fixes : https://github.com/hazelcast/hazelcast/issues/23027

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible

(cherry picked from commit e28b2a1ca8a6d911df453538f8f13ff32ea6819e)