Fix thread interrupt check in SessionAwareSemaphoreReleaseAcquiredSessionsOnFailureTest.testAcquire_shouldReleaseSessionsOnRuntimeError [HZ-1829] (#22995)
The test is failing because AbstractInvocationFuture, which is a custom
implementation of CompletableFuture needs to spin over a loop to detect
if thread is interrupted. However, if semaphore.acquire() response is
received before the loop, the test fails.
Therefore, we need to make sure that semaphore does not respond. I have
changed the semaphore acquire permit from 5 to 6 to make the semaphore
not respond and allow the AbstractInvocationFuture to spin the loop to
do thread.interrupt check

Fixes : https://github.com/hazelcast/hazelcast/issues/20038

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible