Stub schema service methods before the members start in the replication tests
We were using Mockito spies to verify that the members were calling
the replication-related methods of the MemberSchemaService correctly.

However, we were stubbing some of the spied MemberSchemServices
after the members are started.

This is generally working fine, but problematic under the following
scenario, where the same schema service is accessed from different
threads at the same time.

- We call `doThrow(someException).when(schemaService)` in thread A
- At the same time, we call some method of the schema service in
thread B
- The answer (`someException`) is now being associated with the
method called in thread B
- Thread B clears the answers of the invocation container for
the spied schema service, and resets the `stubbingInProgress`
to null
- That write somehow is not visible to thread A, so it continues
with the `.onSchemaAckRequest(any())` part of the stubbing.
- It sees that there are no answers in the invocation container
for that mock, so it prepares to call the real method.
- Before that, it performs some sanity checks to make sure that
there are no stubbing in progress. However, since it can't
see the write done in thread B, it throws.

To avoid this possible scenario, we have to stub the spied
member schema service, before the members start.

This PR refactors the related tests to pass the spies directly
to the factory, where we can stub them beforehand.