Avoid race in JobLifecycleMetricsTest [HZ-1848] (#23055)
The second job submitted intentionally fails. It is possible that the
TerminateExecutionOperation is processed before StartExecutionOperation
reaches this point:

```
ExecutionContext executionContext = executionContexts.get(executionId);
if (executionContext == null) {
    throw new ExecutionNotFoundException(String.format(
	    "%s not found for coordinator %s for '%s'",
	    jobIdAndExecutionId(jobId, executionId), callerAddress, callerOpName));
} else if (!(executionContext.coordinator().equals(callerAddress) && executionContext.jobId() == jobId)) {
```

in `JobExecutionService#assertExecutionContext`

We fix this race by blocking in `MockP#init` before throwing an
exception from it. This way the StartExecutionOperation will run on all
members.