Introduce RepeatOnNodeFailure annotation
The tests that are specifically about the cluster topology, e.g. split-brain tests, should be marked with this annotation, so that the test is repeated upon a node failure. The designated tests are not cancelled as soon as a node fails, which will cause unnecessary waiting when `assertXXXEventually()` methods are used. For early-exit, use `assertNoFailedInstances()`.

Fixes #23078