Fix eventual consistency in WanPublisherMigrationTest.testMigration [HZ-1853] (#23056)
After replica performs MigrationCommitOperation(), it calls the
listeners' onMigrationCommit() method asynchronously. The test was
failing because it was not using eventual consistency to check for the
migration commit counter. I have changed the test to assert the
migration commit counter eventually.

Note : If you want to re-produce the failure before the fix, just add a
sleep(1_000) ms statement to onMigrationCommit() method.

Fixes : https://github.com/hazelcast/hazelcast/issues/22434