Wide log level for PartitionAssignmentsCorrectnessTest [HZ-1864] (#23099)
After a careful logs review of the **only** test failure we may assume
the failure was caused by some environmental reasons, not by the
production code. On other hand, failure is pretty suspicious and we want
to have more information if it fails again. This patch introduces widen
logs for `PartitionAssignmentsCorrectnessTest`.

Closes https://github.com/hazelcast/hazelcast/issues/22122