Fix http timeout in RestLogLevelTest.testDisabledRest [HZ-1865] (#23079)
The test is trying to access a rest endpoint which does not exist. From
the logs we can see that http request blocked on read operation and the
test timed out. When Apache HttpClient is not given any timeout values,
it uses system default values which may not be consistent across all
test environments. So I have changed the test code to use a timeout
value of 20 seconds. 20 seconds is chosen because some tests require a
timeout of at least 10 seconds.

Fixes : https://github.com/hazelcast/hazelcast/issues/21366