Fix some hz-cli quirks on Windows (#22884)
The `hz-cli` shell script fails to run in bash on Windows because it
uses a colon-separated `CLASSPATH` where each path is a linux-style
path, whereas Java on Windows expect a semicolon-separated `CLASSPATH`
where each path is a Windows-style path, even when started from bash.
This PR adds logic to cook the `CLASSPATH` on Windows.

The `hz-cli.bat` cmd script forces the `hazelcast.client.config`
variable in `JAVA_OPTS` which means that if the specified file does not
exist, the command fails. This is annoying when someone wants to specify
an alternate file. This PR removes the line.

In fact, the Windows scripts are lagging behind Bash scripts. This PR
refactors the Windows scripts to align them with what is done with Bash
(same JAVA_OPTS, for instance, and same everything else). Also deals
with `hz-start.bat` via a unique `common.bat` file, same as `common.sh`
for Bash.

Review: should ensure that Bash scripts just still work all right on
Linux, and then that they work on Bash on Windows too, and then that
Windows scripts work too.

Fixes #22883

Checklist:
- [ ] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [ ] Label `Add to Release Notes` or `Not Release Notes content` set
- [ ] Request reviewers if possible
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases
- n/a New public APIs have `@Nonnull/@Nullable` annotations
- n/a New public APIs have `@since` tags in Javadoc