Block communication between instances before closing connections in CompactFormatSplitBrainTest
This test failed a couple of times because the smaller brain
was unable to form a cluster of two members.

The reason was that we were closing the connections between
members (by making them suspect each other) before isolating
both of the clusters.

When I examine the logs from the failed runs, I have seen
that the second member of the smaller cluster was dropping
the first member of the smaller cluster from its member list
and closing its connection to it. When the first member was
trying to claim mastership after we have isolated it from the
larger cluster, the second member was rejecting that claim
because that member was not in its member list, and
we were unable to form the two-member smaller cluster.

That was possible because at the time the first member is
isolated from the larger cluster, the second member
has a perfectly healthy connection to the larger cluster, and
can receive membership changes from the master of the
larger cluster.

As a solution, I block communication between the
larger and smaller clusters, before making them suspect
each other, drop from their member list, and close
connections.

Also, I have cleaned up the test a bit and used
random instances in various steps of the test
to improve it.

Also, I have touched a couple of other split brain
related tests to simplify them.