Wait for safe state in ClientMapLoadAllTest
In the failing run, we can see the following lines in the log.

```
18:46:46,862  INFO |testLoadAll_givenKeys| - [MigrationManager] hz.youthful_diffie.migration - [127.0.0.1]:5701 [dev] [5.2-SNAPSHOT] Repartitioning cluster data. Migration tasks count: 11
18:46:46,877  INFO |testLoadAll_givenKeys| - [MigrationManager] hz.youthful_diffie.migration - [127.0.0.1]:5701 [dev] [5.2-SNAPSHOT] All migration tasks have been completed. (repartitionTime=Tue Aug 02 18:46:46 UTC 2022, plannedMigrations=11, completedMigrations=11, remainingMigrations=0, totalCompletedMigrations=11, elapsedMigrationOperationTime=88ms, totalElapsedMigrationOperationTime=88ms, elapsedDestinationCommitTime=7ms, totalElapsedDestinationCommitTime=7ms, elapsedMigrationTime=97ms, totalElapsedMigrationTime=97ms)
```

It is easy to see that the test can fail when faced with
migrations. When the
`warmUpPartitions(hazelcastFactory.getAllHazelcastInstances());`
line is added to the failing `testLoadAll_givenKeys` test,
the local runs fail with similar exceptions and logs
after some time.

The fix is to make sure that the cluster is safe and there
won't be migrations while the test logic is running.

I have also added a similar logic to the other tests
in this file where we use more than one instance.