Avoid possible NPE while checking TPC ports
In one of the recent PRs, we have removed the client-side
parsing of the TPC-ports, and started using it as it is.

However, we have missed a null check there. And, that
causes NPE when the TPC is disabled, as the ports list
is null.

This PR adds the necessary null check.