Add TPC classes to the URLs of the HazelcastAPIDelegatingClassloader
This is a test fix for the ProcessorClassLoaderTest. In that,
we use the HazelcastAPIDelegatingClassloader, which delegates
all classes not starting with com.hazelcast to system classloader.

For com.hazelcast classes, it performs a lookup on the given URL.
Since Ilogger now depends on TPCLogger, it was not able to find
the TPCLogger class in the hazelcast/target/classes.

Therefore, we need to add the `../hazelcast-tpc-engine/target/classes`
to the URLs, and let it find classes from there as well.