Use Lazy deserialization for queries with nested compact & portable API-1553 (#22360)
DeserializedGenericRecords were being used during queries of nested
compact and portable fields. This behaviour for compact can be observed
by debugging the test `CompactNestedFieldsTest.test_basicQuerying` and
putting a breakpoint in FieldOperations STRING's 
`readGenericRecordOrPrimitive` method. `genericRecord` variable will be
 a `DeserializedGenericRecord`.

I realized it comes from `readAsLeafObjectOnQuery` for the
`FieldKind.COMPACT`. ColumnExpression will return
DeserializedGenericRecord if a compact field is accessed because
 it is the actual column access operation. FieldAccessOperation will
 always return the lazy internal generic record.

If a field is accessed like a.b.c.d 3 FieldAccessOperations will happen
for d, c, and b; and 1 ColumnExpression will happen for a as the last
operation.

Nested querying does not support Array/List or Map access operators in
any form. And there is no array type. So, we don't need to make lazy
deserialization work for that use case and this commit does not include 
the change for it.