[TPC] Backported various improvements from the alto-experimental branch (#23296)
1) AsyncSocket has better to support for dealing with idle connections.
In Java 11 there are additional socket options exposed:
-
https://docs.oracle.com/en/java/javase/11/docs/api/jdk.net/jdk/net/ExtendedSocketOptions.html#TCP_KEEPCOUNT
-
https://docs.oracle.com/en/java/javase/11/docs/api/jdk.net/jdk/net/ExtendedSocketOptions.html#TCP_KEEPIDLE
-
https://docs.oracle.com/en/java/javase/11/docs/api/jdk.net/jdk/net/ExtendedSocketOptions.html#TCP_KEEPINTERVAL
This functionality is exposed on the AsyncSocket. The NIO implementation
relies on reflection and provides an error message when these properties
are not found including suggestions on how to fix the problem. The
advantage of these properties is that you don't need to send bogus
packets to keep the connection alive; the OS takes care of this.

2) AsyncSocket/AsyncServerSocket can be directly created from the
   Eventloop. No need to depend on nio/io_uring after creating the
   eventloop.

3) Added various tests including RPC test. Tests are also set up in such
a way that io_uring can easily be added without copy pasting the tests.

4) ReadHandler is now nio/io_uring agnostic. So after creating the
Nio/io_uring specific configuration, you don't need to know about the
difference between nio/io_uring. It is now fully transparent.

5) Backported some other design changes