[TPC] Another set of changes backported from the alto-branch. (#23366)
Changes:

- The eventloop can get stuck in loop processing tasks from one resource
while ignoring others (so the other resources are starved from
processing time). E.g. when there is a concurrent task that produces a
new concurrent task (and this continues), then the eventloop could get
stuck in processing only concurrent tasks and nothing else. Now there is
a limit on the number of items that can be processed in batch. Once the
maximum number of items in a batch have been reached, the eventloop will
check the other sources of work. Limit on the number of items to process
in batch have been added to the local run queue, concurrent run queue
and scheduled run queue.

- Scheduled tasks do not cause litter due to getting the correct epoch
nanotime.

- There is a cheap nano clock that doesn't call System.nanotime on every
call, but cache the value. This is needed because System.nanotime can be
pretty expensive when the tasks being executed are extremely short.

- The scheduled tasks queue has now an upper limit of tasks. We do not
want to have an uncontrolled number of tasks in the scheduled task
queue.

- Various design improvements: mostly pulling out inner classes.

- Test coverage improvements