[TPC] Socket improvements (#23395)
List of improvements:

- The biggest improvement is the close functionality of the eventloop
and socket/server sockets. There is no more need for the closeableList
for now since the appropriate resources can be acquired from the
SelectionKey. Also, the Selector is now properly closed. So you can just
close the eventloops and the sockets get closed properly without running
into a port that doesn't get released.

- More testing has been added.

- Socket superclass has been pulled out containing the close
functionality

- Socket tracks close cause and reason (like the HZ connections) which
helps with debugging.

- Unit tests run a lot faster. The RPC and large payload test do a much
smaller number of iterations. We do need some nightly tests that run
with a high number of iterations.

- AsyncServerSocket.listen is not needed any longer. AsyncServerSocket
is a passive socket by nature, so it implies that listen needs to be
called. This is now done directly after binding, similar to the
java.nio.ServerSocket. It implies the API and makes it less error-prone.

- AsyncSocket SO_LINGER functionality removed. Its behavior with
non-blocking sockets either is undefined or problematic (on Linux it can
lead to blocking).