[TPC] Introduce ClientAltoConfig
This PR introduces ClientAltoConfig to the ClientConfig.

In it, we only have an enable flag for now. Also, all
the configuration elements and schema definitions are documented
as BETA.

This enabled option will have priority over the setSmartRouting
configuration we have in the ClientNetworkConfig. When Alto config
is enabled, we won't be considering the routing set, always try
to use Alto aware mode.

This PR just adds the configuration and replaces the places
we use routing mode. Further improvments and tests related
to Alto-aware client will be added in a later PR.