Handle null entry (#23439)[5.2.z] (#23450)
Modifications:
- Some stepped-operations were not handling forced-eviction correctly. Forced eviction can remove some 
entries and we can have missing entries in a later step. This was causing NPE in the test linked No way to 
keep the lock when the lock-hold-node was down. #5652
- Returning NULL JsonMetadataStore when MetadataPolicy is OFF. Not returning NULL-meta-data-store was 
causing unexpected HD memory usage hence was breaking the linked test's assumptions: 
com.hazelcast.map.HDMapMemoryLeakStressTest.test_MapOperations hazelcast-enterprise#5652