Use assertTrueEventually in PartitionIndexingTest
The test was failing because not all partitions were being indexed. 101
partitions are expected to be indexed but it was 100. IMO it was because
of the fact that the partition was migrated but the
PartitionMigrationEvent is not received or processed by
MapMigrationAwareService.

So I guess before this line has run for the last time, our assertion
logic ran and indexed partition count was 100 at that moment. I used
`assertTrueEventually` with a 10 seconds of timeout where we use
`assertPartitionsIndexedCorrectly` to fix the flakiness.

There won't be a ee couterpart of this PR. Looks like the test is added
in 3.12 and the issue exists there as well. So, we will port the fix to
4.1.z and above. (since 4.0.z is out of support). I will send those when
the PR is approved.

backport of #23179