Use assertTrueEventually in PartitionIndexingTest API-1666 [4.2.z] (#23458)
The test was failing because not all partitions were being indexed. 101
partitions are expected to be indexed but it was 100. IMO it was because
of the fact that the partition was migrated but the
PartitionMigrationEvent is not received or processed by
MapMigrationAwareService.

So I guess before this line has run for the last time, our assertion
logic ran and indexed partition count was 100 at that moment. I used
`assertTrueEventually` with a 10 seconds of timeout where we use
`assertPartitionsIndexedCorrectly` to fix the flakiness.

backport of #23179