Reactor renaming
The Eventloop has been renamed to Reactor. And also
EventloopBuilder has been renamed to ReactorBuilder.

The actual Eventloop logic has been pulled out into a class
named Eventloop. Many of the methods of the Reactor have been
moved into the Eventloop. It makes it a lot easier to separate
the state of the Eventloop from the Reactor.

Unsafe has been merged into the Eventloop.