Reactor renaming (#23510)
The Eventloop has been renamed to Reactor. And also EventloopBuilder has
been renamed to ReactorBuilder.

The actual Eventloop logic has been pulled out into a class named
Eventloop. Many of the methods of the Reactor have been moved into the
Eventloop. It makes it a lot easier to separate the state of the
Eventloop from the Reactor and it will be easier to fix false sharing
problems.

Unsafe has been merged into the Eventloop.

We don't need to change the configuration of eventloops in Alto.
Configuring the eventloop count still applies since the concept of
eventloop remains.