Refactor SqlConnector API (#23443)
In preparation for https://hazelcast.atlassian.net/browse/HZ-1526

Some connectors needed `Expression` and some `RexNode`. For this we had
two overloads, one with both types of objects, and one with just
`Expression`, for b-w compatibility.

This PR uses just `RexNode`, and provides a `context` object the
connector can use to convert `RexNode` to `Expression`. We also moved
the `DAG` and `Table` objects to the context, to reduce the number of
parameters. In the future we can add more context objects without
breaking the compatibility.

Breaking the compatibility now is allowed because the `SqlConnector` is
still not a public API.

---------

Co-authored-by: Krzysztof Jamróz <79092062+k-jamroz@users.noreply.github.com>