NioAsyncSocket process OP_WRITE before OP_READ.
When OP_READ is called, typically it leads to an increase in memory usage
because all kinds of data is loaded from socket buffers into memory.

By calling OP_WRITE first, typically it leads to a reduction in because
we move things from in memory into the socket buffer.

So when you do OP_READ first, you first increase and then reduce. But
when by doing OP_WRITE first, you first reduce and then increease.