[5.2.z] Allow serializing an unsupported type with reflective serializer if there's an explicit serializer for it API-1829  (#23569)
Backport of: #23449 

This is requested from a customer through
salesforce
https://hazelcast.lightning.force.com/lightning/r/Case/5006e00001rNjuwAAC/view

There is a PR opened https://github.com/hazelcast/hazelcast/pull/23372
specifically for `java.time.Instant`. We decided to solve this
generally, so in this commit, I allowed serializing unsupported types
via zero config serializer if there's a explicit serializer for the
unsupported type.