Redesign of AsyncSocket.activate
The idea behind AsyncSocket.activate was that you could create a
an AsyncSocket on one reactor and then activate it on another. This
is useful on the serverside so you can run the accepted socket on
a different reactor than the AsyncServerSocket.

This lead to a bit of a weird design because you need to active with a
reactor you want to attach to and made it impossible to set fields of
the AsyncSocket on creation.

This PR solves that by always making the AsyncSocket on the appropriate
Reactor. And instead of listening to created AsyncSocket on the
AsyncServerSocket, you listen to AcceptRequest and then you create the
Socket on the appropriate reactor.